requirements.txt
flask
torch
ultralytics # for yolov5
transformers
Pillow
pytesseract
TTS
gunicorn
opencv-python